Test: TeamMate Battle

+++ src/main/java/dungeonmania/entities/movingEntities/ZombieToastEntity.java
@@ -34,13 +34,13 @@ public class ZombieToastEntity extends Entity implements IContactingEntity, IBat
         this.health = health;
     }
 
-    public int getDamage() {
+    public float getDamage() {
         // TODO determine correct ZombieToast damage
         return 3;
     }
 
     @Override
-    public void loseHealth(float enemyHealth, int enemyDamage) {
+    public void loseHealth(float enemyHealth, float enemyDamage) {
         this.health -= ((enemyHealth * enemyDamage) / 5);
     }
 //endregion
